ULTIMA8: Move getExtraGuiOption() into UltimaMetaEngineDetection
