+++ engines/ultima/ultima8/meta_engine.cpp
@@ -124,79 +124,6 @@ static const KeybindingRecord DEBUG_KEYS[] = {
 };
 #endif
 
-static const ExtraGuiOption COMMON_OPTIONS[] = {
-	{
-		_s("Enable frame skipping"),
-		_s("Allow the game to skip animation frames when running too slow."),
-		"frameSkip",
-		false
-	},
-	{
-		_s("Enable frame limiting"),
-		_s("Limits the speed of the game to prevent running too fast."),
-		"frameLimit",
-		true
-	},
-	{
-		_s("Enable cheats"),
-		_s("Allow cheats by commands and a menu when player is clicked."),
-		"cheat",
-		false
-	},
-	{
-		_s("Enable high resolution"),
-		_s("Enable a higher resolution for the game"),
-		"usehighres",
-		false
-	},
-	{ nullptr, nullptr, nullptr, false }
-};
-
-static const ExtraGuiOption U8_OPTIONS[] = {
-	{
-		_s("Play foot step sounds"),
-		_s("Plays sound when the player moves."),
-		"footsteps",
-		true
-	},
-	{
-		_s("Enable jump to mouse position"),
-		_s("Jumping while not moving targets the mouse cursor rather than direction only."),
-		"targetedjump",
-		true
-	},
-	{
-		_s("Use original save/load screens"),
-		_s("Use the original save/load screens instead of the ScummVM ones"),
-		"originalsaveload",
-		false
-	},
-	{
-		_s("Enable font replacement"),
-		_s("Replaces game fonts with rendered fonts"),
-		"font_override",
-		false
-	},
-	{
-		_s("Enable font anti-aliasing"),
-		_s("When font anti-aliasing is enabled, the text is smoother."),
-		"font_antialiasing",
-		false
-	},
-	{ nullptr, nullptr, nullptr, false }
-};
-
-static const ExtraGuiOption CRUSADER_OPTIONS[] = {
-	{
-		// I18N: Silencer is the player-character in Crusader games, known as the Avatar in Ultima series.
-		_s("Camera moves with Silencer"),
-		_s("Camera tracks the player movement rather than snapping to defined positions."),
-		"camera_on_player",
-		true
-	},
-	{ nullptr, nullptr, nullptr, false }
-};
-
 Common::KeymapArray MetaEngine::initKeymaps(const Common::String &gameId, bool isMenuActive) {
 	Common::KeymapArray keymapArray;
 
@@ -282,22 +209,6 @@ Common::KeymapArray MetaEngine::initKeymaps(const Common::String &gameId, bool i
 #endif
 
 	return keymapArray;
-}
-
-const ExtraGuiOptions MetaEngine::getExtraGuiOptions(const Common::String& target) {
-	ExtraGuiOptions options;
-
-	for (const ExtraGuiOption *o = COMMON_OPTIONS; o->configOption; ++o) {
-		options.push_back(*o);
-	}
-
-	// Game specific options
-	const ExtraGuiOption *game_options = (target.equals("ultima8") ? U8_OPTIONS : CRUSADER_OPTIONS);
-	for (const ExtraGuiOption *o = game_options; o->configOption; ++o) {
-		options.push_back(*o);
-	}
-
-	return options;
 }
 
 void MetaEngine::setTextInputActive(bool isActive) {
